hz_langgraph_learning/agent-dev/requirements.txt
# Core LangGraph and LangChain
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langsmith>=0.1.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async HTTP and Web Scraping
aiohttp>=3.9.0
playwright>=1.40.0

# Environment and Configuration
python-dotenv>=1.0.0

# Utilities
python-multipart>=0.0.6
